Abstracted the overlay :art:
